pyproject.toml
[tool.poetry]
name = "strawberry-graphql"
packages = [ { include = "strawberry" } ]
version = "0.10.0"
description = "A library for creating GraphQL APIs"
authors = ["Patrick Arminio <patrick.arminio@gmail.com>"]
license = "MIT"
readme = "README.md"
keywords = ["grapqhl", "api", "rest", "starlette", "async"]

homepage = "https://strawberry.rocks/"
repository = "https://github.com/strawberry-graphql/strawberry"
documentation = "https://strawberry.rocks/"

classifiers = [
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = "^3.7"
graphql-core-next = "^1.0"
starlette = "=0.12"
click = "^7.0"
hupper = "^1.5"
pygments = "^2.3"
uvicorn = "^0.4.6"

[tool.poetry.dev-dependencies]
pytest = "^4.5"
pdbpp = "^0.10.0"
bpython = "^0.18"
pytest-emoji = "^0.2.0"
flake8 = "^3.7"
black = {version = "^19.3b0",allows-prereleases = true}
isort = "^4.3"
pytest-asyncio = "^0.10.0"

[tool.poetry.scripts]
strawberry = "strawberry.cli:run"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
